// SDKs/amazonaws.com/cognito-idp/2016-04-18/typescript/src/sdk/models/operations/updateuserpoolclient.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import * as shared from "../shared";

export enum UpdateUserPoolClientXAmzTargetEnum {
    AwsCognitoIdentityProviderServiceUpdateUserPoolClient = "AWSCognitoIdentityProviderService.UpdateUserPoolClient"
}


export class UpdateUserPoolClientHeaders extends SpeakeasyBase {
  @Metadata({ data: "header, style=simple;explode=false;name=X-Amz-Algorithm" })
  xAmzAlgorithm?: string;

  @Metadata({ data: "header, style=simple;explode=false;name=X-Amz-Content-Sha256" })
  xAmzContentSha256?: string;

  @Metadata({ data: "header, style=simple;explode=false;name=X-Amz-Credential" })
  xAmzCredential?: string;

  @Metadata({ data: "header, style=simple;explode=false;name=X-Amz-Date" })
  xAmzDate?: string;

  @Metadata({ data: "header, style=simple;explode=false;name=X-Amz-Security-Token" })
  xAmzSecurityToken?: string;

  @Metadata({ data: "header, style=simple;explode=false;name=X-Amz-Signature" })
  xAmzSignature?: string;

  @Metadata({ data: "header, style=simple;explode=false;name=X-Amz-SignedHeaders" })
  xAmzSignedHeaders?: string;

  @Metadata({ data: "header, style=simple;explode=false;name=X-Amz-Target" })
  xAmzTarget: UpdateUserPoolClientXAmzTargetEnum;
}


export class UpdateUserPoolClientRequest extends SpeakeasyBase {
  @Metadata()
  headers: UpdateUserPoolClientHeaders;

  @Metadata({ data: "request, media_type=application/json" })
  request: shared.UpdateUserPoolClientRequest;
}


export class UpdateUserPoolClientResponse extends SpeakeasyBase {
  @Metadata()
  concurrentModificationException?: any;

  @Metadata()
  contentType: string;

  @Metadata()
  internalErrorException?: any;

  @Metadata()
  invalidOAuthFlowException?: any;

  @Metadata()
  invalidParameterException?: any;

  @Metadata()
  notAuthorizedException?: any;

  @Metadata()
  resourceNotFoundException?: any;

  @Metadata()
  scopeDoesNotExistException?: any;

  @Metadata()
  statusCode: number;

  @Metadata()
  tooManyRequestsException?: any;

  @Metadata()
  updateUserPoolClientResponse?: shared.UpdateUserPoolClientResponse;
}
